renamed file to compact_test.dart
